Add Docker Usage to Readme
This commit adds instructions for building and running the Swift Jupyter Kernel using the Docker container defined by the Dockerfile in this repo.

The required run parameters for the container weren't completely obvious, and it took some trial and error to figure out what was needed to make the container usable, so this added documentation attempts to fill in those gaps.

The addition of the security option `--security-opt seccomp:unconfined` was taken from this thread: 

https://groups.google.com/a/tensorflow.org/forum/#!msg/swift/ru0g4A29D8c/fcOeFGhIAAAJ

As discussed there, this option allows the container to make system calls on the host, which is less than ideal.  It does fix a fatal error which occurred when the kernel is run inside the container without this setting, but if there is another way to solve this it would probably be preferred.