Allow package install location to be set.
The preferred package install location can be set using the
%install-location magic command, which points to the destination
directory for compiled Swift modules. This allows the user to compile
and install packages in a notebook (using the %install directives), and
then import them in a different notebook without recompiling.

The compiled packages are built as dependencies of a
`jupyterInstalledPackages` target. Note that compilation generates a
dylib with that target, that must be `dlopen`ed before pre-built
packages can be imported. For this reason, the `.so` file is also
copied inside the package install location, and lazily opened when the
Swift process is started.

Limitations:
- Build artifacts are not cached. When restarting a notebook, comment
or skip installation directives (but keep `%install-location`) to
prevent the recompilation of packages previously built.