Share build artifacts in addition to built modules.
The Jupyter notebook library target will be rebuilt every time. However,
SPM properly determines the dependencies that need to be rebuilt, so no
unnecessary download / compilation steps should be necessary.

Regenerating the target is much better for consistency, and ensures the
notebook does not rely on external state.

In order to implement this change, an additional container directory
has been created, and that container is symlinked to the user-specified
location when appropriate.