Modulemap directories get the name of the package (#61)
The name of the package is used instead of a numeric suffix, so we get
directories like these:
 modulemap-COpenCV/
 modulemap-CSwiftVips/
 modulemap-opencv4/
 modulemap-vips/

Instead of modulemap-0, etc.

The reason for doing this is that the modulemap-<number> directories
sometimes got polluted after installing packages across different
notebooks. This is made worse by the fact that `swift package
show-dependencies` sometimes declares dependencies for packages in the
.build directory that were left out by previous installations, and are
not really in use for the package that is being currently compiled.

After installing multiple packages, we could end up in a situation
where multiple modulemap- directories refer to the same package,
because one dependency was assigned slot '1' in one installation and
'0' in another one. When this happens, the process ended up in a "bad
state" and we had to resort to removing the swift-install folder and
start all over again.

Another alternative would have been to delete all modulemap directories
when installing a new package. But I have something else coming in a
future pull request that will take advantage of reusing all installed
modules.