Support load from Bazel
This added support to load Swift libraries from Bazel. It makes
assumption that Jupyter is launched from the root of Bazel. I am still
figuring out how subprocess.check_output bazel query can be slower.

How to:

```
%bazel "//nnc:nnc"
%bazel "@PythonKit//:PythonKit"
```

This will install Bazel packages available locally, and made them
available for the notebook to use.

There is a more detailed discussion of steps we do to achieve this.